+++ core_cv/comment/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Comment
+
+admin.site.register(Comment)
