+++ core_cv/comment/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CommentConfig(AppConfig):
+    name = 'comment'
